Backend/requirements-docker.txt
fastapi
uvicorn[standard]
python-dotenv
supabase==2.15.0
httpx>=0.24.0,<0.29.0
groq
pypdf
numpy==1.26.4
torch==2.3.1
sentence-transformers==3.0.1
transformers==4.44.2
faiss-cpu==1.8.0
scikit-learn
scipy
huggingface-hub
pydantic
requests
python-multipart
flask
python-dateutil
aiofiles
tokenizers==0.19.1